requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==33.0.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
